svn merge -c 37326 https://source.sakaiproject.org/svn/assignment/trunk U    assignment-bundles/assignment.properties U    assignment-tool/tool/src/java/org/sakaiproject/assignment/tool/AssignmentAction.java U    assignment-impl/impl/src/java/org/sakaiproject/assignment/impl/BaseAssignmentService.java svn log -r 37326 https://source.sakaiproject.org/svn/assignment/trunk ------------------------------------------------------------------------ r37326 | zqian@umich.edu | 2007-10-23 12:16:57 -0400 (Tue, 23 Oct 2007) | 1 line
fix to SAK-11992:Deleted assignment appearing to students but not instructor; when the assignment is non-electronic type, delete assignment also removes all submissions and assignment content object. Otherwise, if there is no submission to the assignment yet, both assignment and assignment content objects will be removed. If there is submission object, assignment is just marked as 'deleted'
------------------------------------------------------------------------


git-svn-id: https://source.sakaiproject.org/svn/assignment/branches/sakai_2-5-x@37543 66ffb92e-73f9-0310-93c1-f5514f145a0a